+++ astropy/coordinates/tests/test_distance.py
@@ -14,10 +14,10 @@ tests.  Right now that's just regression tests.
 
 
 def test_distance_change():
-    from .. import RA, Dec, ICRSCoordinates, Distance
+    from .. import Longitude, Latitude, ICRSCoordinates, Distance
 
-    ra = RA("4:08:15.162342", unit=u.hour)
-    dec = Dec("-41:08:15.162342", unit=u.degree)
+    ra = Longitude("4:08:15.162342", unit=u.hour)
+    dec = Latitude("-41:08:15.162342", unit=u.degree)
     c = ICRSCoordinates(ra, dec)
 
     c.distance = Distance(1, unit=u.kpc)
@@ -31,10 +31,10 @@ def test_distance_change():
     assert c.x == oldx * 2
 
 def test_distance_from_quantity():
-    from .. import RA, Dec, ICRSCoordinates, Distance
+    from .. import Longitude, Latitude, ICRSCoordinates, Distance
 
-    ra = RA("4:08:15.162342", unit=u.hour)
-    dec = Dec("-41:08:15.162342", unit=u.degree)
+    ra = Longitude("4:08:15.162342", unit=u.hour)
+    dec = Latitude("-41:08:15.162342", unit=u.degree)
     c = ICRSCoordinates(ra, dec)
 
     # a Quantity object should be able to supply a distance